# Bool writer like reader